wip: code adjustment and optimization
