+++ src/App.vue
@@ -9,13 +9,12 @@
 <script lang="ts">
   import { defineComponent } from 'vue';
   import { ConfigProvider } from 'ant-design-vue';
+  import { AppProvider } from '/@/components/Application';
 
   import { initAppConfigStore } from '/@/logics/initAppConfig';
 
   import { useLockPage } from '/@/hooks/web/useLockPage';
   import { useLocale } from '/@/locales/useLocale';
-
-  import { AppProvider } from '/@/components/Application';
 
   export default defineComponent({
     name: 'App',
