+++ src/hooks/event/useScrollTo.ts
@@ -1,6 +1,5 @@
 import { isFunction, isUnDef } from '/@/utils/is';
 import { ref, unref } from 'vue';
-import { requestAnimationFrame } from '/@/utils/animation';
 
 export interface ScrollToParams {
   el: HTMLElement;
