+++ src/utils/http/axios/Axios.ts
@@ -81,7 +81,9 @@ export class VAxios {
     // 请求拦截器配置处理
     this.axiosInstance.interceptors.request.use((config: AxiosRequestConfig) => {
       // If cancel repeat request is turned on, then cancel repeat request is prohibited
-      const { headers: { ignoreCancelToken } = { ignoreCancelToken: false } } = config;
+      const {
+        headers: { ignoreCancelToken = false },
+      } = config;
       !ignoreCancelToken && axiosCanceler.addPending(config);
       if (requestInterceptors && isFunction(requestInterceptors)) {
         config = requestInterceptors(config);
