+++ src/views/sys/error-log/index.vue
@@ -1,9 +1,9 @@
 <template>
   <div class="p-4">
-    <template v-for="src in imgListRef" :key="src">
+    <template v-for="src in imgList" :key="src">
       <img :src="src" v-show="false" />
     </template>
-    <DetailModal :info="rowInfoRef" @register="registerModal" />
+    <DetailModal :info="rowInfo" @register="registerModal" />
     <BasicTable @register="register" class="error-handle-table">
       <template #toolbar>
         <a-button @click="fireVueError" type="primary">
@@ -50,8 +50,8 @@
     name: 'ErrorHandler',
     components: { DetailModal, BasicTable, TableAction },
     setup() {
-      const rowInfoRef = ref<ErrorInfo>();
-      const imgListRef = ref<string[]>([]);
+      const rowInfo = ref<ErrorInfo>();
+      const imgList = ref<string[]>([]);
 
       const { t } = useI18n();
 
@@ -84,7 +84,7 @@
       }
       // 查看详情
       function handleDetail(row: ErrorInfo) {
-        rowInfoRef.value = row;
+        rowInfo.value = row;
         openModal(true);
       }
 
@@ -93,7 +93,7 @@
       }
 
       function fireResourceError() {
-        imgListRef.value.push(`${new Date().getTime()}.png`);
+        imgList.value.push(`${new Date().getTime()}.png`);
       }
 
       async function fireAjaxError() {
@@ -107,8 +107,8 @@
         fireVueError,
         fireResourceError,
         fireAjaxError,
-        imgListRef,
-        rowInfoRef,
+        imgList,
+        rowInfo,
         t,
       };
     },
