--- src/views/sys/exception/exception.less
@@ -1,5 +1,0 @@
-.exception {
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-}
