+++ src/views/sys/exception/index.ts
@@ -1,1 +1,1 @@
-export { default as Exception } from './Exception';
+export { default as Exception } from './Exception.vue';
