+++ src/views/sys/iframe/index.vue
@@ -1,31 +1,44 @@
 <template>
-  <div class="iframe-page" :style="getWrapStyle">
+  <div :class="prefixCls" :style="getWrapStyle">
     <Spin :spinning="loading" size="large" :style="getWrapStyle">
-      <iframe :src="frameSrc" class="iframe-page__main" ref="frameRef"></iframe>
+      <iframe :src="frameSrc" :class="`${prefixCls}__main`" ref="frameRef"></iframe>
     </Spin>
   </div>
 </template>
 <script lang="ts">
-  import { defineComponent, PropType, ref, unref, onMounted, nextTick, computed } from 'vue';
+  import type { CSSProperties } from 'vue';
+  import { defineComponent, ref, unref, onMounted, nextTick, computed } from 'vue';
   import { Spin } from 'ant-design-vue';
 
   import { getViewportOffset } from '/@/utils/domUtils';
+
   import { useWindowSizeFn } from '/@/hooks/event/useWindowSizeFn';
+
+  import { propTypes } from '/@/utils/propTypes';
+  import { useDesign } from '/@/hooks/web/useDesign';
 
   export default defineComponent({
     name: 'IFrame',
     components: { Spin },
     props: {
-      frameSrc: {
-        type: String as PropType<string>,
-        default: '',
-      },
+      frameSrc: propTypes.string.def(''),
     },
     setup() {
-      const loadingRef = ref(false);
+      const loading = ref(false);
       const topRef = ref(50);
       const heightRef = ref(window.innerHeight);
-      const frameRef = ref<HTMLElement | null>(null);
+      const frameRef = ref<HTMLFrameElement | null>(null);
+
+      const { prefixCls } = useDesign('iframe-page');
+      useWindowSizeFn(calcHeight, 150, { immediate: true });
+
+      const getWrapStyle = computed(
+        (): CSSProperties => {
+          return {
+            height: `${unref(heightRef)}px`,
+          };
+        }
+      );
 
       function calcHeight() {
         const iframe = unref(frameRef);
@@ -40,21 +53,19 @@
         iframe.style.height = `${clientHeight}px`;
       }
 
-      useWindowSizeFn(calcHeight, 150, { immediate: true });
-
       function hideLoading() {
-        loadingRef.value = false;
+        loading.value = false;
         calcHeight();
       }
 
       function init() {
         nextTick(() => {
           const iframe = unref(frameRef);
-          if (!iframe) {
-            return;
-          }
-          if ((iframe as any).attachEvent) {
-            (iframe as any).attachEvent('onload', () => {
+          if (!iframe) return;
+
+          const _frame = iframe as any;
+          if (_frame.attachEvent) {
+            _frame.attachEvent('onload', () => {
               hideLoading();
             });
           } else {
@@ -65,23 +76,23 @@
         });
       }
       onMounted(() => {
-        loadingRef.value = true;
+        loading.value = true;
         init();
       });
+
       return {
-        getWrapStyle: computed(() => {
-          return {
-            height: `${unref(heightRef)}px`,
-          };
-        }),
-        loading: loadingRef,
+        getWrapStyle,
+        loading,
         frameRef,
+        prefixCls,
       };
     },
   });
 </script>
 <style lang="less" scoped>
-  .iframe-page {
+  @prefix-cls: ~'@{namespace}-iframe-page';
+
+  .@{prefix-cls} {
     .ant-spin-nested-loading {
       position: relative;
       height: 100%;
