+++ src/views/sys/redirect/index.vue
@@ -12,7 +12,9 @@
 
       const { params, query } = unref(currentRoute);
       const { path } = params;
+
       const _path = Array.isArray(path) ? path.join('/') : path;
+
       replace({
         path: '/' + _path,
         query,
